Workaround for native.keyboardshow called twice
When switching focus between inputs, the event native.keyboardshow is called twice. On the second event, the dataset ionkeyboard.bottom of the element .content.overflow-scroll, is set to the already changed bottom value (the keyboard height). After, when the keyboard hides, the bottom value is set to the keyboard height and the view shows a blank space where the keyboard was.